Initial Creation
Powershell script to get file size and size on disk of all files in a directory, as well as their file path. This only gets files with the attributes "Archive, SparseFile, ReparsePoint, Offline".
  
No other file will be found. A quick modification of removing " -and $_.Attributes -eq 'Archive, SparseFile, ReparsePoint, Offline'" on line 82 can remove this. This is useful for when finding Commvault stubbed files.   

Originally created and taken from this post below, but modified to meet my requirements: 
  https://www.opentechguides.com/how-to/article/powershell/133/size-on-disk-ps.html